blog-creator/requirements.txt
python-dotenv
supabase
requests
openai
anthropic
langfuse
fastapi
uvicorn